Added Features:
Working Ratings screen
Settings panel WIA

What's Included in Settings panel
logout :>

+++ app/src/main/java/com/example/mobdeveapplication/datasets/Historyobject.kt
@@ -1,4 +1,4 @@
 package com.example.mobdeveapplication.datasets
 
-data class Historyobject(val location:String,val rating:Int) {
+data class Historyobject(val location:String, val rating: Float) {
 }
